feat: Add basic page skeleton structure
We need to start adding styles to the main page. To make it easier to
implement the main navigation, we can add some basic structure that
would help us visualize the card list.

This commit adds the bare bones of the main page structure with the
proper font and sizes.