feat: Implement game list with data
We want to show the list of games in the main page and ensure that it's
properly shown. We want a grid-like layout with consistent sizes for
the cards.

This commit adds the parts needed to render the grid and to size the
elements as we want to ensure that the cards' design don't break.