feat: Implement add and remove to cart logic
We want to allow users add games to the cart and remove them.
We also want to store the games in the localstorage to ensure that if
the user refreshes the page, the cart keeps its state.

This commit adds the logic to let the user add and remove games by
clicking the button. It also changes the label for games that are
already in the cart. We keep the state in the localstorage as well.