feat: Add endpoint and action to fetch by id
We are storing the ids in localstorage and not the whole item to avoid
duplicating the source of truth for games and ending up with stale
data.
The games endpoint is paginated which means that we would need to fetch
every game to restore the cart only with the ids.
Since we store the items in the localstorage, we cannot access the ids
in the server without passing them directly from the client.

This commit creates a way to fetch the games by ids reusing the same
endpoint that we had.