feat: Extract cart logic into custom hook
We need to reuse logic that handles removing items from the cart in the
cart page. We also need to access the data in the cart within that
page.

This commit extracts that logic into a custom hook to reuse it when
needed. Since we are using localstorage to store the games, we can
safely reuse the hook without the accidentally re-instantiating the
same hook over and over again.